Move 'login' package outside of graphql clients pkg
