+++ src/cmd/login/login.go
@@ -2,9 +2,9 @@ package login
 
 import (
 	"context"
-	"github.com/otterize/otterize-cli/src/pkg/cloudclient/graphql/login/auth_api"
-	"github.com/otterize/otterize-cli/src/pkg/cloudclient/graphql/login/server"
 	"github.com/otterize/otterize-cli/src/pkg/cloudclient/graphql/users"
+	"github.com/otterize/otterize-cli/src/pkg/cloudclient/login/auth_api"
+	"github.com/otterize/otterize-cli/src/pkg/cloudclient/login/server"
 	cloudclient "github.com/otterize/otterize-cli/src/pkg/cloudclient/restapi"
 	"github.com/otterize/otterize-cli/src/pkg/cloudclient/restapi/cloudapi"
 	"github.com/otterize/otterize-cli/src/pkg/config"
