+++ src/pkg/cloudclient/login/authenticator/auth.go
