+++ src/pkg/cloudclient/login/server/server.go
@@ -3,8 +3,8 @@ package server
 import (
 	"encoding/base64"
 	"fmt"
-	"github.com/otterize/otterize-cli/src/pkg/cloudclient/graphql/login/auth_api"
-	"github.com/otterize/otterize-cli/src/pkg/cloudclient/graphql/login/authenticator"
+	"github.com/otterize/otterize-cli/src/pkg/cloudclient/login/auth_api"
+	"github.com/otterize/otterize-cli/src/pkg/cloudclient/login/authenticator"
 	"github.com/otterize/otterize-cli/src/pkg/utils/must"
 	"github.com/sirupsen/logrus"
 	"golang.org/x/exp/rand"
